% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneable_hdoutliers.R
\docType{data}
\name{knn_search_type}
\alias{knn_search_type}
\alias{values_knn_search_type}
\title{Nearest Neighbour Search Methods}
\format{
An object of class \code{character} of length 3.
}
\usage{
knn_search_type(values = values_knn_search_type)

values_knn_search_type
}
\arguments{
\item{values}{A character string of possible values. See \code{values_knn_search_type}
in examples below.}
}
\description{
Nearest Neighbour Search Methods
}
\details{
This parameter is used in \code{straystep::step_hdoutliers}.
}
\examples{
values_knn_search_type
knn_search_type()
}
\keyword{datasets}
