% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_hdoutliers.R
\name{step_hdoutliers}
\alias{step_hdoutliers}
\title{XXXX}
\usage{
step_hdoutliers(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  reference_colnames = NULL,
  outlier_bounds = NULL,
  outlier_cutoff_threshold = 0.01,
  k_neighbours = 10,
  knnsearchtype = "brute",
  normalize_method = "unitize",
  candidate_proportion = 0.5,
  threshold_sample_size = 50,
  options = list(),
  skip = FALSE,
  id = recipes::rand_id("hdout")
)
}
\arguments{
\item{...}{One or more selector functions to choose which
variables will be used to compute the components. See
\code{\link[=selections]{selections()}} for more details. For the \code{tidy}
method, these are not currently used.}

\item{role}{For model terms created by this step, what analysis
role should they be assigned?. By default, the function assumes
that the new principal component columns created by the original
variables will be used as predictors in a model.}

\item{options}{A list of options to the default method for
\link{XXXX}.}

\item{prefix}{A character string that will be the prefix to the
resulting new variables.}

\item{ref_dist}{placeholder}
}
\value{
An updated version of \code{recipe} with the new step
added to the sequence of existing steps (if any). For the
\code{tidy} method, a tibble with columns \code{terms} (the
selectors or variables selected), \code{value} (the
loading), and \code{component}.
}
\description{
\code{XXX} creates a \emph{specification} of a recipe step
that will ....
}
\details{
Some text here
}
\examples{
x <- 1
}
\references{
Add reference
}
\concept{dwt}
\concept{preprocessing}
\concept{projection_methods}
\keyword{datagen}
